% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inProgress.R
\name{all_BSJ_to_circRNA_sequence}
\alias{all_BSJ_to_circRNA_sequence}
\title{all_BSJ_to_circRNA_sequence}
\usage{
all_BSJ_to_circRNA_sequence(BSJ_df, genome, TxDb, annotationLibrary)
}
\arguments{
\item{BSJ_df}{: Dataframe containing BSJ coordinates and annotated geneID}

\item{genome}{: Is the length f the library fragment}

\item{TxDb}{: The sequence read length}
}
\value{
Returns a DNAstring object.
}
\description{
Takes an annotated df that contains BSJ coordinates and
returns the predicted circular RNA sequence.
}
\examples{

library('Ularcirc')
TxDb <- TxDb.Hsapiens.UCSC.hg38.knownGene::TxDb.Hsapiens.UCSC.hg38.knownGene
genome <- BSgenome.Hsapiens.UCSC.hg38::BSgenome.Hsapiens.UCSC.hg38
BSJ  <- 'chr2_40430305_chr2_40428472'   # chr2_40430302_chr2_40428472
annotationLibrary <- org.Hs.eg.db
BSJ_to_circRNA_sequence(BSJ, "SLC8A1", genome,TxDb, annotationLibrary)
# chr11_33286412_chr11_33287512

TxDb_hg19 <- TxDb.Hsapiens.UCSC.hg19.knownGene::TxDb.Hsapiens.UCSC.hg19.knownGene
genome_hg19 <- BSgenome.Hsapiens.UCSC.hg19::BSgenome.Hsapiens.UCSC.hg19


}
