% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ularcirc_standalone.R
\name{BSJ_details}
\alias{BSJ_details}
\title{BSJ_details
This function returns details of a BSJ string and returns a list of coordinates. Can accept two
different formats, Ularcirc or  generic.}
\usage{
BSJ_details(BSJ)
}
\arguments{
\item{BSJ}{: backsplice junction as a string. See details below for example formats}
}
\description{
BSJ_details
This function returns details of a BSJ string and returns a list of coordinates. Can accept two
different formats, Ularcirc or  generic.
}
\examples{
bsj <-  'chr14_99465814_chr14_99458278'   # Historic Ularcirc format

bsj <- c("chr14_99465814_chr14_99458278","chr22_20933778_chr22_20934245",
         "chr12_120155720_chr12_120154969", "chr4_143543508_chr4_143543973",
         "chr10_7285955_chr10_7276891")

BSJ_details(bsj)

bsj <- 'chr10:100923974-100926020:+'     # generic format
BSJ_details(bsj)

}
