% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ularcirc_standalone.R
\name{bsj_fastq_generate}
\alias{bsj_fastq_generate}
\title{bsj_fastq_generate}
\usage{
bsj_fastq_generate(circRNA_Sequence, fragmentLength = 300,
  readLength = 100, variations = 4, headerID = "")
}
\arguments{
\item{circRNA_Sequence}{: Linear sequence of a circRNA. i.e. the backspice junction
is the first and last base of this sequence}

\item{fragmentLength}{: Is the length f the library fragment}

\item{readLength}{: The sequence read length}
\item{variations}{: Number of sequences returned for each read type. Note each sequence variation will start at a unique location (where possible)}
\item{headID}{: Character identifier that will be incorporated into sequence header }
}
\value{
Returns a list of two DNAstring sets labelled "read1" and "read2" which correspond to forward and reverse read pairs.
}
\description{
Takes a circRNA predicted sequence and generates synthetic sequence reads that
}
\examples{

library('Ularcirc')


# Obtain a circRNA sequence
circRNA_Sequence <- "GGAAGAGGAAGAACGTCTGAGAAATAAAATTCGAGCTGATCATGAGAAGGCCTTGGAAGAAGCAAAAGAAAAATTAAGAAAGTCAAGAGAGGAAATTCGAGCAGAAATTCAGACAGAGAAAAATAAGGTAGTCCAAGAAATGAAGATAAAAGAGAACAAGCCACTGCCACCAGTCCCTATTCCCAACCTTGTAGGAATACGTGGTGGAGACCCAGAAGATAATGACATAAGAGAGAAAAGGGAAAAAATTAAAGAGATGATGAAACATGCTTGGGATAACTATAGGACATATGGGTGGGGACATAATGAACTCAGACCTATTGCAAGGAAAGGACACTCCCCTAACATATTTGGAAGTTCACAAATGGGTGCTACCATAGTAGATGCTTTGGATACCCTTTATATCATGGGACTTCATGATGAATTCCTAGATGGGCAAAGATGGATTGAAGACAACCTTGATTTCAGTGTGAATTCAGAGGTGTCTGTGTTTGAAGTCAACATTCGATTTATTGGAGGCCTACTTGCAGCATATTACCTATCAGGAGAGGAG"
fastqReads <- bsj_fastq_generate(circRNA_Sequence, fragmentLength=300, readLength=100, variations = 4, headerID='circRNA_example')
# The following will indicate 12 sequences are present in each list entry
length(fastqReads$read1)
length(fastqReads$read2)

# can create fastq files as follows
Biostrings::writeXStringSet(fastqReads$Read_One,"circRNA_Sample_R1.fastq.gz",compress = TRUE, format="fastq")
Biostrings::writeXStringSet( fastqReads$Read_Two,"circRNA_Sample_R2.fastq.gz",compress = TRUE, format="fastq")


}
