% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inProgress.R
\name{BSJ_relative_to_Exons}
\alias{BSJ_relative_to_Exons}
\title{library('Ularcirc')
TxDb <- TxDb.Hsapiens.UCSC.hg38.knownGene::TxDb.Hsapiens.UCSC.hg38.knownGene
BSJ  <- 'chr2_40430305_chr2_40428472'   # chr2_40430302_chr2_40428472
BSJ  <- 'chr14_21269406_chr14_21234076'; geneID <- 'HNRNPC'
annotationLibrary <- org.Hs.eg.db::org.Hs.eg.db}
\usage{
BSJ_relative_to_Exons(BSJ, geneID = NULL, TxDb, annotationLibrary,
  strand = "*")
}
\description{
library('Ularcirc')
TxDb <- TxDb.Hsapiens.UCSC.hg38.knownGene::TxDb.Hsapiens.UCSC.hg38.knownGene
BSJ  <- 'chr2_40430305_chr2_40428472'   # chr2_40430302_chr2_40428472
BSJ  <- 'chr14_21269406_chr14_21234076'; geneID <- 'HNRNPC'
annotationLibrary <- org.Hs.eg.db::org.Hs.eg.db
}
