% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_io.R
\name{chimericFilters}
\alias{chimericFilters}
\title{chimericFilters}
\usage{
chimericFilters(
  BSjuncName = NULL,
  sortDir = "Descending",
  indexNumber = 1,
  displayNumber = 10,
  displayRADscore = FALSE,
  RADcountThreshold = 10,
  applyFSJfilter = FALSE
)
}
\arguments{
\item{BSjuncName}{: A character string that represents a backsplice junction ID. Set when
needing to extract a specific junction. Default NULL.}

\item{sortDir}{: Specifies how data is sorted, either "Descending" (default) or "Ascending".}

\item{indexNumber}{: Filter data according to this file index}

\item{displayNumber}{: Number of records to display in an shiny app}

\item{displayRADscore}{: Boolean. If TRUE then will apply/calculate RAD score}

\item{RADcountThreshold}{: Integer. The minimum count threshold required to calculate RAD score.
i.e. A default RAD score of -1 will be applied to any BSJ with a count less than this score}

\item{applyFSJfilter}{: Boolean of whether to apply FSJ filter}
}
\description{
This function sets the basic filters that are
}
