% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ularcirc_miRNA.R
\name{circSequence}
\alias{circSequence}
\title{circSequence
This function returns pre-defined full length circRNA sequences.}
\usage{
circSequence(circRNA_ID = "slc8a1")
}
\arguments{
\item{circRNA_ID}{: Character string of a circRNA_ID}
}
\description{
circSequence
This function returns pre-defined full length circRNA sequences.
}
\examples{
      circSequence(circRNA_ID="slc8a1")


}
