% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ularcirc.R
\name{Compatible_Annotation_DBs}
\alias{Compatible_Annotation_DBs}
\title{Compatible_Annotation_DBs}
\usage{
Compatible_Annotation_DBs(list_commands = FALSE)
}
\arguments{
\item{list_commands}{Boolean. By default this is FALSE and will return a dataframe of all
compatible databases. When set to TRUE will return a list of commands that can be copied
onto console which will download the appropriate databases (see example below).}
}
\value{
Returns a list of installation commands for all compatible annotation databases
}
\description{
Interogates bioconductor databases and identifies those that are compatible
with Ularcirc. Builds a list of commands that the user can copy to install the
required database on their local computer. Once installed the databases are
immediately available to Ularcirc upon re-starting the shiny app.
This function requires connection to the internet.
}
\examples{
# Get all bioconductor annotation databases that are compatible with Ularcirc
library('BSgenome')
library('httpuv')
library('AnnotationHub')
# Prepare a dataframe of all compatible annotation databases
all_db_df <- Compatible_Annotation_DBs()

# Prepare a list of R commands that users can copy paste to download
# and install annotation databases on local machine
db_download_commands <- Compatible_Annotation_DBs(list_commands = TRUE)

# Select a database and display the commands needed to install
# Use noquote to correctly format output
\donttest{ noquote(db_download_commands$Hsapiens.UCSC.hg38) }
}
