% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_io.R
\name{FilterChimericJuncs}
\alias{FilterChimericJuncs}
\title{FilterChimeric}
\usage{
FilterChimericJuncs(
  All_junctions,
  chromFilter = TRUE,
  strandFilter = TRUE,
  genomicDistance = c(200, 1e+05),
  canonicalJuncs = TRUE,
  fileID = c(-1),
  chrM_Filter = TRUE,
  invertReads = FALSE
)
}
\arguments{
\item{fileID}{: blah balh}

\item{ChromFilter}{: when TRUE (default) both chimera parts have to align to same chromosome}

\item{StrandFilter}{: when TRUE (default) both chimera parts have to align to same strand}

\item{Genomic_Distance}{: minimum and maximum distance filters of chimeric reads on chromosome. Only is applied
if ChromFilter is TRUE and StrandFilter is TRUE}

\item{CanonicalJuncs}{: when TRUE}

\item{ChrM_Filter}{: blah blah}

\item{InvertReads}{: blah blah}
}
\description{
A generic function that filters STAR chimeric junction files on certain genomic criteria (eg strand, same chromosome etc).
Useful filter to remove the most obvious false positives. The default filter settings are suitable for circRNA
 discovery in humans / mice data sets.
}
\seealso{
SelectUniqueJunctions, loadSTAR_chimeric
}
