% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ularcirc_miRNA.R
\name{load_miRBase_fasta}
\alias{load_miRBase_fasta}
\title{load_miRBase_fasta
This function extracts species specific mirbase sequences as a Biostrings RNAstringset.}
\usage{
load_miRBase_fasta(
  mirbaselocalDir = tempdir(),
  speciesCode = "hsa",
  filename = "hairpin.fa"
)
}
\arguments{
\item{mirbaselocalDir}{: Directory where miRbase files are located. If files are not detected will download requested files from mirBase}

\item{speciesCode}{: Three letter code of species (eg human = hsa). Used to subset species miRbase entries}

\item{filename}{: The miRBase filename. Expects this to be a fasta formatted file. If file does not exist will download from mirBase.}
}
\description{
load_miRBase_fasta
This function extracts species specific mirbase sequences as a Biostrings RNAstringset.
}
