% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_io.R
\name{SelectUniqueJunctions}
\alias{SelectUniqueJunctions}
\title{SelectUniqueJunctions}
\usage{
SelectUniqueJunctions(
  BSJ_junctions,
  filterlist = chimericFilters(),
  library_strand = "Opposing strand",
  FSJ_Junctions = NULL,
  shinyapp = FALSE
)
}
\arguments{
\item{BSJ_junctions}{: Junction to display}

\item{filterlist}{: filterlist}

\item{library_strand}{: default is "Opposing strand"}

\item{FSJ_Junctions}{: Junction to display.}

\item{shinyapp}{: Boolean. If true used to setup control status bars in shiny app.}
}
\description{
Builds a summary table from chimeric data obtained from the STAR aligner. Assembles table with
the requested number of top entries. Populates with RAD score and FSJ score.
}
\details{
This is the workhorse for collated BSJ junctions from the input data. It will return
   selected rows of data (annotated) that will enable enhanced browsing of raw data on the fly.

	Filter options: Junction abundance. Sort
   library_strand = "Opposing strand" or "Same strand" or "Unstranded"
}
