% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ularcirc_standalone.R
\name{sequence_from_exon_coords}
\alias{sequence_from_exon_coords}
\title{sequence_from_exon_coords}
\usage{
sequence_from_exon_coords(genome, exon_df)
}
\arguments{
\item{genome}{: genome object}

\item{exon_df}{: data frame of exons. Must have column with names "chrom", "start",
"stop", "strand"}
}
\description{
sequence_from_exon_coords
}
