% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ularcirc.R
\name{Ularcirc}
\alias{Ularcirc}
\title{Ularcirc}
\usage{
Ularcirc()
}
\value{
Does not return anything
}
\description{
When the function is invoked the Ularcirc shiny app is started. The starting screen has quickstart instructions on how to use the software.
Please refer to the Ularcirc vignette for a more detailed workflow.
}
\examples{
# The following commands will load the shiny app either through an RStudio session or
# through your internet browser
\dontrun{
library("Ularcirc")
Ularcirc()
}
}
