% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ularcirc_standalone.R, R/inProgress.R
\name{BSJ_to_circRNA_sequence}
\alias{BSJ_to_circRNA_sequence}
\alias{BSJ_to_circRNA_sequence}
\title{BSJ_to_circRNA_sequence}
\usage{
BSJ_to_circRNA_sequence(BSJ, geneID = NULL, genome, TxDb, annotationLibrary)

BSJ_to_circRNA_sequence(BSJ, geneID = NULL, genome, TxDb, annotationLibrary)
}
\arguments{
\item{BSJ}{: BSJ coordinate}

\item{genome}{: Is the length f the library fragment}

\item{TxDb}{: The sequence read length}

\item{BSJ}{: BSJ coordinate}

\item{genome}{: Is the length f the library fragment}

\item{TxDb}{: The sequence read length}
}
\value{
Returns a DNAstring object.

Returns a DNAstring object.
}
\description{
Takes a BSJ coordinates and generates a predicted circular RNA sequence

Takes a BSJ coordinates and generates a predicted circular RNA sequence
}
\examples{

library('Ularcirc')
TxDb <- TxDb.Hsapiens.UCSC.hg38.knownGene::TxDb.Hsapiens.UCSC.hg38.knownGene
genome <- BSgenome.Hsapiens.UCSC.hg38::BSgenome.Hsapiens.UCSC.hg38
BSJ  <- 'chr2_40430305_chr2_40428472'   # chr2_40430302_chr2_40428472
annotationLibrary <- org.Hs.eg.db
BSJ_to_circRNA_sequence(BSJ, "SLC8A1", genome,TxDb, annotationLibrary)
# chr11_33286412_chr11_33287512

TxDb_hg19 <- TxDb.Hsapiens.UCSC.hg19.knownGene::TxDb.Hsapiens.UCSC.hg19.knownGene
genome_hg19 <- BSgenome.Hsapiens.UCSC.hg19::BSgenome.Hsapiens.UCSC.hg19


library('Ularcirc')
TxDb <- TxDb.Hsapiens.UCSC.hg38.knownGene::TxDb.Hsapiens.UCSC.hg38.knownGene
genome <- BSgenome.Hsapiens.UCSC.hg38::BSgenome.Hsapiens.UCSC.hg38
BSJ  <- 'chr2_40430305_chr2_40428472'   # chr2_40430302_chr2_40428472
annotationLibrary <- org.Hs.eg.db
BSJ_to_circRNA_sequence(BSJ, "SLC8A1", genome,TxDb, annotationLibrary)
# chr11_33286412_chr11_33287512

TxDb_hg19 <- TxDb.Hsapiens.UCSC.hg19.knownGene::TxDb.Hsapiens.UCSC.hg19.knownGene
genome_hg19 <- BSgenome.Hsapiens.UCSC.hg19::BSgenome.Hsapiens.UCSC.hg19


}
