% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ularcirc_standalone.R
\name{bsj_to_circRNA_sequence}
\alias{bsj_to_circRNA_sequence}
\title{bsj_to_circRNA_sequence}
\usage{
bsj_to_circRNA_sequence(BSJ, geneID = NULL, genome, TxDb,
  annotationLibrary)
}
\arguments{
\item{BSJ}{: BSJ coordinate in the format of chr_coordinate_chr_coorindate OR chr:coordinate-coorindate:strand.}

\item{geneID}{: The gene ID that the BSJ aligns to. Not essential as this can
be identified from the BSJ coordinate, however time performance of function improved
if this information can be provided.}

\item{genome}{: Is the length f the library fragment}

\item{TxDb}{: The sequence read length}

\item{annotationLibrary}{: annotation database. See details for example.}
}
\value{
Returns a DNAstring object.
}
\description{
Takes one BSJ coordinate and generates a predicted circular RNA sequence.
}
\examples{

library('Ularcirc')
library('BSgenome.Hsapiens.UCSC.hg38')
library('TxDb.Hsapiens.UCSC.hg38.knownGene')
TxDb <- TxDb.Hsapiens.UCSC.hg38.knownGene::TxDb.Hsapiens.UCSC.hg38.knownGene
genome <- BSgenome.Hsapiens.UCSC.hg38::BSgenome.Hsapiens.UCSC.hg38
annotationLibrary <- org.Hs.eg.db::org.Hs.eg.db

# Define BSJ. Following two formats are accepted
BSJ <- 'chr2:40430305-40428472:-'       # SLC8A1
BSJ  <- 'chr2_40430305_chr2_40428472'   # SLC8A1

circRNA_sequence <- bsj_to_circRNA_sequence(BSJ, "SLC8A1", genome,TxDb, annotationLibrary)

# You can also retrieve sequence without passing gene annotation - but this is slower
# circRNA_sequence <- bsj_to_circRNA_sequence(BSJ, NULL, genome,TxDb, annotationLibrary)


}
