% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_io.R
\name{chimericStats}
\alias{chimericStats}
\title{chimericStats}
\usage{
chimericStats(chimericDT)
}
\arguments{
\item{chimericDT}{: Data table of chimeric junctions as provided by
STAR aligner}
}
\description{
Number of chimeric entries
  Number of unique junctions
  Frequency of most abundant junction
  Number of backsplice vs other junctions
}
