% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ularcirc.R
\name{Compatible_Annotation_DBs}
\alias{Compatible_Annotation_DBs}
\title{Compatible_Annotation_DBs}
\usage{
Compatible_Annotation_DBs()
}
\value{
Returns a list of installation commands for all compatible annotation databases
}
\description{
Interogates bioconductor databases and identifies those that are compatible
with Ularcirc. Builds a list of commands that the user can copy to install the
required database on their local computer. Once installed the databases are
immediately available to Ularcirc upon re-starting the shiny app.
This function requires connection to the internet.
}
\examples{
# Get all bioconductor annotation databases that are compatible with Ularcirc
library('BSgenome')
library('httpuv')
library('AnnotationHub')
all_dbs <- Compatible_Annotation_DBs()

# List database IDs
names(all_dbs)

# Select a database and display the commands needed to install
# Use noquote to correctly format output
\donttest{noquote(all_dbs$Hsapiens.UCSC.hg38) }
}
