% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_io.R
\name{FilterChimeric_Ularcirc}
\alias{FilterChimeric_Ularcirc}
\title{Wrapper function for Ularcirc shiny app which expects a list of objects to be returned}
\usage{
FilterChimeric_Ularcirc(
  All_junctions,
  ChromFilter = TRUE,
  StrandFilter = TRUE,
  Genomic_Distance = c(200, 1e+05),
  CanonicalJuncs = TRUE,
  fileID = c(-1),
  ChrM_Filter = TRUE,
  InvertReads = FALSE,
  SummaryNumber = 50
)
}
\description{
Wrapper function for Ularcirc shiny app which expects a list of objects to be returned
}
